moved home screen code to new file
