+++ economicalc_client/lib/main.dart
@@ -2,6 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+import 'package:economicalc_client/home_screen.dart';
 import 'package:flutter/material.dart';
 
 void main() {
@@ -9,65 +10,13 @@ void main() {
 }
 
 class MyApp extends StatelessWidget {
-
-  Widget titleSection=Container(
-    child: Column(
-      mainAxisAlignment: MainAxisAlignment.spaceAround,
-      children: <Widget>[
-        Padding(padding: const EdgeInsets.all(8.0),
-        child:Text(
-          "EconomiCalc",
-          style: TextStyle(
-            color: new Color(0xff000000),
-            fontWeight: FontWeight.bold,
-            fontSize: 36.0
-          )))
-      ],
-    ),
-  );
-
-
-  static GestureDetector buildButtonColumn(IconData icon) {
-  
-  Color color = Colors.black;
-  return GestureDetector(
-    onTap: (){print("hej");},
-    child: Column(
-    mainAxisSize: MainAxisSize.min,
-    mainAxisAlignment: MainAxisAlignment.center,
-    children: <Widget>[
-      Padding(
-        padding: EdgeInsets.all(10.0),
-      ),
-      Icon(icon, color: color),
-    ],
-  )
-  );
-}
-  Widget iconSection=Container(
-    child: Row(mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-    children: <Widget>[
-      buildButtonColumn(Icons.scanner),
-      buildButtonColumn(Icons.search),
-      buildButtonColumn(Icons.auto_graph),
-      buildButtonColumn(Icons.filter)
-      //TODO replace with figma svgs
-    ]),
-  );
+  const MyApp({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'EconomiCalc',
-      theme: new ThemeData(scaffoldBackgroundColor: const Color(0xFFB8D8D8)),
-      home: Scaffold (
-        body: ListView(
-          children: <Widget>[
-            titleSection,
-            iconSection,
-          ],
-          )
-        ),
-    );
+        title: 'EconomiCalc',
+        theme: new ThemeData(scaffoldBackgroundColor: const Color(0xFFB8D8D8)),
+        home: HomeScreen());
   }
 }
